app/container_builds/event_processor_for_svc_bus_queues/requirements.txt
Flask
gunicorn

azure-functions
azure-identity
azure-appconfiguration
azure-storage-blob
azure-servicebus
azure-cosmos
azure-eventhub
azure-eventhub-checkpointstoreblob
azure-monitor-opentelemetry

asyncio
aiohttp

azure-eventhub-checkpointstoreblob-aio
azure-identity

azure-core-tracing-opentelemetry
azure-monitor-opentelemetry-exporter
